// lib/shared/types.ts
    interface Response {
      error: (code: number, error: any) => void
    }
// ...
export enum ApprovalStates {
  Accepted = 'Accepted',
  Declined = 'Declined',
  NoResponse = 'No Response',
}

export enum ApprovalCategory {
  Upload = 'Upload',
  Deployment = 'Deployment',
}

export interface DeploymentMetadata {
  highLevelDetails: {
    name: string
    [x: string]: unknown
  }

  contacts: {
    owner: Array<Entity>
    [x: string]: unknown
  }
}
// ...
export interface LogStatement {
  timestamp: Date
  level: string
  msg: string
}

export type ModelId = string | Types.ObjectId

export enum SchemaType {
  UPLOAD = 'UPLOAD',
  DEPLOYMENT = 'DEPLOYMENT',
}

export interface Schema {
  name: string
  reference: string
  schema: any
  use: SchemaType
}

// Dates are in ISO 8601 format
enum DateStringBrand {}
export type DateString = string & DateStringBrand

export enum EntityKind {
  USER = 'user',
}

export interface Entity {
  kind: EntityKind
  id: string
}
// ...
export interface Deployment {
  _id: any
  schemaRef: string | null
  uuid: string

  model: Types.ObjectId | ModelDoc
  metadata: DeploymentMetadata

  managerApproved: ApprovalStates

  logs: Types.Array<LogStatement>
  built: boolean
  ungoverned: boolean

  createdAt: Date
  updatedAt: Date

  log: (level: string, msg: string) => Promise<void>
}

export type DeploymentDoc = Deployment & Document<any, any, Deployment>

export interface Model {
  _id: any
  schemaRef: string
  uuid: string

  versions: Types.Array<VersionDoc | Types.ObjectId>
  latestVersion: VersionDoc | Types.ObjectId

  createdAt: Date
  updatedAt: Date
}

export type ModelDoc = Model & Document<any, any, Model>